requirements.txt.txt
streamlit
streamlit_option_menu
os
ABC
abstractmethod
PIL
pandas
matplotlib.pyplot